feat: Migrate fetchVaultTvl to utils/fetchPrice.js
